Secrets can not be passed via outputs
So we add the ACR_USER and ACR_PW to the github secrets and move the Key Vault action to the ACI Deploy job. 

If we don't add ACR_USER and ACR_PW to github secrets, we'd have to access the key vault multiple times due to different jobs needing the secrets. Seems wasteful